Kosmetika kódu - S1118 (abstract class with public constructor) a IDE0004 (redundant cast)

+++ Havit.Data/Trace/DbCommandTraceData.cs
@@ -91,7 +91,7 @@ namespace Havit.Data.Trace
 					sb.AppendLine($"    {parameter.ParameterName}: {parameter.Value} (DbType.{parameter.DbType}, {parameter.Direction})");
 				}
 			}
-			decimal durationMs = (decimal)this.DurationTicks / (decimal)TimeSpan.TicksPerMillisecond;
+			decimal durationMs = this.DurationTicks / (decimal)TimeSpan.TicksPerMillisecond;
             sb.AppendLine($"  Duration: {durationMs:N2} ms");
 			
 			if (ResultSet)
